requirements.txt
python-espncricinfo
espn-api
celery
beautifulsoup4